\name{rundssat}

\alias{rundssat}

\title{Individual function to run DSSAT}

\usage{
rundssat(i, path.to.extdata, TRT, AOI)
}

\arguments{
\item{i}{run number}
\item{TRT}{is the number of treatments to run from the experimental file}
}

\value{
DSSAT outputs
}

\description{
Individual function to run DSSAT
}

\examples{
rundssat(1)
}
