% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{SimSim-package}
\alias{SimSim}
\alias{SimSim-package}
\title{Similarity between Similarity Matrices}
\description{
To be provided
}
\references{
To follow
}
\author{
\strong{Maintainer}: Daniel Lawson \email{dan.lawson@bristol.ac.uk}

}
