\name{slu1}

\alias{slu1}

\title{Evaporation limit function from Ritchie et al. (1989); cited in Allen et al. (2005)}

\usage{
slu1(clay1, sand1)
}

\arguments{
\item{clay1}{Clay percentage for the top soil horizon}
\item{sand1}{Sand percentage for the top soil horizon}
}

\description{
Evaporation limit function from Ritchie et al. (1989); cited in Allen et al. (2005)
}

\keyword{internal}
